gdxVersion = 1.9.13

+++ core/src/com/betmansmall/game/gameInterface/InterfaceSelector.java
@@ -358,7 +358,7 @@ public class InterfaceSelector extends Table {
         return flinging;
     }
 
-    public boolean scrolled(int amount) {
+    public boolean scrolled(float amount) {
         int x = gameInterface.prevMouseX;
         int y = gameInterface.prevMouseY;
         if (vertical) {
